Get the artworks information within a classification

Get the artworks information within a classification

get_artwork_info(classification = "Prints", size = 10, token = token)

Arguments

classification
An art classification according to Harvard Art Museum website. The full list of available classification is: 'Prints', 'Albums', 'Amulets', 'Armor', 'Boxes', 'Calligraphy', 'Cameos', 'Fragments', 'Furnishings', 'Gems', 'Inscriptions', 'Jewelry', 'Mirrors', 'Mosaics', 'Paintings', 'Photographs', 'Plaques', 'Rubbings', 'Sculpture', 'Seals', 'Tablets', 'Tokens', 'Vessels'. Default is set to be 'Prints'.

size
Size specifies how many pieces of artwork the user would like to get. Default is set to be 10.

token
A variable from token <- get_token() function. Please run token <- get_token() before using this function

Value

a list of artwork information under a specific classification. The information includes title, century, period, culture, division, dclassification, department, objectid, corresponding url and imageurl.

Examples

get_artwork_info(classification='Prints', size=10)

#> please get your token thorough `token <- get_token()` and input token=token in this function : )